catalog/core/catalog-core-commands/src/main/java/org/codice/ddf/commands/util/ValidatorConfigReader.java
/**
 * Copyright (c) Codice Foundation
 * <p>
 * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser
 * General Public License as published by the Free Software Foundation, either version 3 of the
 * License, or any later version.
 * <p>
 * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
 * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * Lesser General Public License for more details. A copy of the GNU Lesser General Public License
 * is distributed along with this program and can be found at
 * <http://www.gnu.org/licenses/lgpl.html>.
 */

package org.codice.ddf.commands.util;

import java.io.InputStream;

import javax.xml.bind.JAXBContext;
import javax.xml.bind.JAXBException;
import javax.xml.bind.Unmarshaller;

public class ValidatorConfigReader {

    private RuleSet ruleSet;

    /**
     * Constructor
     *
     * @param
     */
    public ValidatorConfigReader(InputStream inputStream) throws JAXBException {
        JAXBContext jaxbContext = JAXBContext.newInstance(RuleSet.class);
        Unmarshaller unmarshaller = jaxbContext.createUnmarshaller();
        ruleSet = (RuleSet) unmarshaller.unmarshal(inputStream);
    }

    /**
     * Returns a java object which represents the scoring rubric for the validator(s)
     *
     * @return
     */
    public RuleSet getRuleSet() {
        return ruleSet;
    }

    public void setRuleSet(RuleSet ruleSet) {
        this.ruleSet = ruleSet;
    }
}
